edited config example to add one with quotes
I was having issues with the component failing to load without any useful error messages, and with some trial and error I found that putting the game in quotes helped. I'm guessing it has to do with the colon. Added it as an example to show to users/readers they may need to put it in quotes